feat: Create project python API
