+++ skore/tests/integration/cli/test_create.py
@@ -1,67 +1,6 @@
 import subprocess
 
 import pytest
-from skore.cli.create_project import (
-    InvalidProjectNameError,
-    ProjectAlreadyExistsError,
-    ProjectCreationError,
-    __create,
-    validate_project_name,
-)
-
-test_cases = [
-    (
-        "a" * 250,
-        (False, InvalidProjectNameError()),
-    ),
-    (
-        "%",
-        (False, InvalidProjectNameError()),
-    ),
-    (
-        "hello world",
-        (False, InvalidProjectNameError()),
-    ),
-]
-
-
-@pytest.mark.parametrize("project_name,expected", test_cases)
-def test_validate_project_name(project_name, expected):
-    result, exception = validate_project_name(project_name)
-    expected_result, expected_exception = expected
-    assert result == expected_result
-    assert type(exception) is type(expected_exception)
-
-
-@pytest.mark.parametrize("project_name", ["hello", "hello.skore"])
-def test_create_project(project_name, tmp_path):
-    __create(project_name, working_dir=tmp_path)
-    assert (tmp_path / "hello.skore").exists()
-
-
-# TODO: If using fixtures in test cases is possible, join this with
-# `test_create_project`
-def test_create_project_absolute_path(tmp_path):
-    __create(tmp_path / "hello")
-    assert (tmp_path / "hello.skore").exists()
-
-
-def test_create_project_fails_if_file_exists(tmp_path):
-    __create(tmp_path / "hello")
-    assert (tmp_path / "hello.skore").exists()
-    with pytest.raises(ProjectAlreadyExistsError):
-        __create(tmp_path / "hello")
-
-
-def test_create_project_fails_if_permission_denied(tmp_path):
-    with pytest.raises(ProjectCreationError):
-        __create("/")
-
-
-@pytest.mark.parametrize("project_name", ["hello.txt", "%%%", "COM1"])
-def test_create_project_fails_if_invalid_name(project_name, tmp_path):
-    with pytest.raises(ProjectCreationError):
-        __create(project_name, working_dir=tmp_path)
 
 
 def test_create_project_cli_default_argument(tmp_path):
